Fiddle with selection
There was a bigger change here, but the approach turned out to
have some problems with large sets.  The current app saves and
restores the selected rows when you make an edit, retaining the set
of selected bytes even if the number of lines changes (maybe you
reformatted bytes into a string).  There's no way to do that quickly
with WPF when the number of selected items gets large (say 10K+).
I will probably just cap the selection, and refuse to restore it if
it exceeds a certain size.

The ListView SelectedItems management seems to use an O(n^2) (or
worse) algorithm.  It might be trying to verify that items being
added to SelectedItems actually exist in Items -- I can see it
calling Contains().  Whatever the case, it's a big step backward
performance-wise from WinForms.  Yay WPF.

See the DisasmUiTest project's Selection Test to see what I tried.