Wire up sub-window double-click handlers and list nav
Wired up the double-click handlers for References, Notes, and
Symbols.  These jump to the item at the offset that was double-
clicked.  Also hooked up the navigate forward/backward buttons
in the toolbar.

Except for the usual WPF gymnastics required to figure out what you
actually double-clicked on, this went smoothly.