Set Owner for all modal dialogs
This turns out to be really important.  Otherwise the modal dialog
doesn't stay on top of the application's window stack, which can
make things awkward when ShowInTaskbar is set to false.  The easiest
way to ensure this is getting done is to make it part of the
constructor arguments.

The code now passes the parent window in explicitly.  WPF MessageBox
avoids this by calling UnsafeNativeMethods.GetActiveWindow(), but
that feels weird.  We could assume Application.Current.MainWindow
is the parent, but that seems like it could go quietly and horribly
wrong.