Finish Edit Data Operand dialog
I looked at doing this in full WPF fashion, binding context object
properties to radio button IsChecked properties, but there's a lot
of conditional logic behind the scenes.  I also wanted to do the
various formatted strings with a StringFormat binding, but that only
takes one argument, and some of them need two.

On the plus side, the format strings are now in Window.Resources,
rather than the global string table or the buttons themselves.
(The latter always felt a little sleazy, but it seemed like the
least annoying way to do it in WinForms.)