Fix obscure list traversal issue
If you select a note, delete it, select some nearby lines, and hit
undo, then the next time you hit up/down arrow the list will jump
back to the first line.  The workaround for this appears to be to
set the focus on a selected ListViewItem from an obscure event.

Maybe there's a simpler way that I just missed?  This is absurd.
(Yay WPF.)